+++ crates/stackable-webhook/src/servers/conversion.rs
@@ -78,6 +78,12 @@ pub struct ConversionWebhookOptions {
     /// The name of the Kubernetes service which points to the operator/webhook.
     pub service_name: String,
 
+    /// If the CRDs should be maintained automatically. Use the (negated) value from
+    /// `stackable_operator::cli::ProductOperatorRun::disable_crd_maintenance`
+    /// for this.
+    // # Because of https://github.com/rust-lang/cargo/issues/3475 we can not use a real link here
+    pub maintain_crds: bool,
+
     /// The field manager used to apply Kubernetes objects, typically the operator name, e.g.
     /// `airflow-operator`.
     pub field_manager: String,
@@ -97,11 +103,12 @@ impl ConversionWebhookServer {
     /// Creates a new conversion webhook server, which expects POST requests being made to the
     /// `/convert/{crd name}` endpoint.
     ///
-    /// You need to provide two things for every CRD passed in via the `crds_and_handlers` argument:
+    /// You need to provide a few things for every CRD passed in via the `crds_and_handlers` argument:
     ///
     /// 1. The CRD
     /// 2. A conversion function to convert between CRD versions. Typically you would use the
-    ///   the auto-generated `try_convert` function on CRD spec definition structs for this.
+    ///    the auto-generated `try_convert` function on CRD spec definition structs for this.
+    /// 3. A [`kube::Client`] used to create/update the CRDs.
     ///
     /// The [`ConversionWebhookServer`] takes care of reconciling the CRDs into the Kubernetes
     /// cluster and takes care of adding itself as conversion webhook. This includes TLS
@@ -119,14 +126,18 @@ impl ConversionWebhookServer {
     /// use stackable_operator::{
     ///     kube::Client,
     ///     crd::s3::{S3Connection, S3ConnectionVersion},
-    ///     cli::OperatorEnvironmentOptions,
+    ///     cli::ProductOperatorRun,
     /// };
     ///
     /// # async fn test() {
     /// // Things that should already be in you operator:
     /// const OPERATOR_NAME: &str = "product-operator";
     /// let client = Client::try_default().await.expect("failed to create Kubernetes client");
-    /// let operator_environment = OperatorEnvironmentOptions::parse();
+    /// let ProductOperatorRun {
+    ///     operator_environment,
+    ///     disable_crd_maintenance,
+    ///     ..
+    /// } = ProductOperatorRun::parse();
     ///
     ///  let crds_and_handlers = [
     ///     (
@@ -140,9 +151,10 @@ impl ConversionWebhookServer {
     ///     socket_addr: format!("0.0.0.0:{CONVERSION_WEBHOOK_HTTPS_PORT}")
     ///         .parse()
     ///         .expect("static address is always valid"),
-    ///     field_manager: OPERATOR_NAME.to_owned(),
     ///     namespace: operator_environment.operator_namespace,
     ///     service_name: operator_environment.operator_service_name,
+    ///     maintain_crds: !disable_crd_maintenance,
+    ///     field_manager: OPERATOR_NAME.to_owned(),
     /// };
     ///
     /// // Construct the conversion webhook server
@@ -205,9 +217,10 @@ impl ConversionWebhookServer {
 
         let ConversionWebhookOptions {
             socket_addr,
-            field_manager,
             namespace: operator_namespace,
             service_name: operator_service_name,
+            maintain_crds,
+            field_manager,
         } = &options;
 
         // This is how Kubernetes calls us, so it decides about the naming.
@@ -233,28 +246,33 @@ impl ConversionWebhookServer {
             .recv()
             .await
             .context(ReceiveCertificateFromChannelSnafu)?;
-        Self::reconcile_crds(
-            &client,
-            field_manager,
-            &crds,
-            operator_namespace,
-            operator_service_name,
-            current_cert,
-        )
-        .await
-        .context(ReconcileCrdsSnafu)?;
 
-        try_join!(
-            Self::run_webhook_server(server),
-            Self::run_crd_reconciliation_loop(
-                cert_rx,
+        if *maintain_crds {
+            Self::reconcile_crds(
                 &client,
                 field_manager,
                 &crds,
                 operator_namespace,
                 operator_service_name,
-            ),
-        )?;
+                current_cert,
+            )
+            .await
+            .context(ReconcileCrdsSnafu)?;
+
+            try_join!(
+                Self::run_webhook_server(server),
+                Self::run_crd_reconciliation_loop(
+                    cert_rx,
+                    &client,
+                    field_manager,
+                    &crds,
+                    operator_namespace,
+                    operator_service_name,
+                ),
+            )?;
+        } else {
+            Self::run_webhook_server(server).await?;
+        };
 
         Ok(())
     }