+++ crates/stackable-operator/src/config/merge.rs
@@ -6,8 +6,7 @@ use std::{
 };
 
 use k8s_openapi::{
-    DeepMerge,
-    api::core::v1::{NodeAffinity, PodAffinity, PodAntiAffinity, PodTemplateSpec},
+    api::core::v1::{NodeAffinity, PodAffinity, PodAntiAffinity},
     apimachinery::pkg::{api::resource::Quantity, apis::meta::v1::LabelSelector},
 };
 pub use stackable_operator_derive::Merge;
@@ -87,11 +86,6 @@ impl<K: Hash + Eq + Clone, V: Merge + Clone> Merge for HashMap<K, V> {
         }
     }
 }
-impl Merge for PodTemplateSpec {
-    fn merge(&mut self, defaults: &Self) {
-        self.merge_from(defaults.clone())
-    }
-}
 
 /// Moving version of [`Merge::merge`], to produce slightly nicer test output
 pub fn merge<T: Merge>(mut overrides: T, defaults: &T) -> T {
@@ -162,8 +156,6 @@ impl<T: Atomic> Merge for Option<T> {
 #[cfg(test)]
 mod tests {
     use std::collections::{BTreeMap, HashMap};
-
-    use k8s_openapi::api::core::v1::{PodSpec, PodTemplateSpec};
 
     use super::{Merge, merge};
 
@@ -435,61 +427,6 @@ mod tests {
                 &[("a", Acc(3)), ("c", Acc(5))].into()
             ),
             BTreeMap::from([("a", Acc(4)), ("b", Acc(2)), ("c", Acc(5))])
-        );
-    }
-
-    #[test]
-    fn merge_pod_template_spec() {
-        assert_eq!(
-            merge(
-                PodTemplateSpec {
-                    spec: Some(PodSpec {
-                        service_account_name: Some("my-sa".to_string()),
-                        ..PodSpec::default()
-                    }),
-                    ..PodTemplateSpec::default()
-                },
-                &PodTemplateSpec {
-                    spec: Some(PodSpec {
-                        priority: Some(3000),
-                        ..PodSpec::default()
-                    }),
-                    ..PodTemplateSpec::default()
-                }
-            ),
-            PodTemplateSpec {
-                spec: Some(PodSpec {
-                    service_account_name: Some("my-sa".to_string()),
-                    priority: Some(3000),
-                    ..PodSpec::default()
-                }),
-                ..PodTemplateSpec::default()
-            }
-        );
-        assert_eq!(
-            merge(
-                PodTemplateSpec {
-                    spec: Some(PodSpec {
-                        service_account_name: Some("sa-to-be-overridden".to_string()),
-                        ..PodSpec::default()
-                    }),
-                    ..PodTemplateSpec::default()
-                },
-                &PodTemplateSpec {
-                    spec: Some(PodSpec {
-                        service_account_name: Some("sa-override".to_string()),
-                        ..PodSpec::default()
-                    }),
-                    ..PodTemplateSpec::default()
-                }
-            ),
-            PodTemplateSpec {
-                spec: Some(PodSpec {
-                    service_account_name: Some("sa-override".to_string()),
-                    ..PodSpec::default()
-                }),
-                ..PodTemplateSpec::default()
-            }
         );
     }
 }