+++ crates/stackable-operator/CHANGELOG.md
@@ -6,9 +6,9 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- BREAKING: Add a new CLI flag/env to disabling CRD maintenance: `--disable-crd-maintenance` ([#10XX]).
+- BREAKING: Add a new CLI flag/env to disabling CRD maintenance: `--disable-crd-maintenance` ([#1085]).
 
-[#10XX]: https://github.com/stackabletech/operator-rs/pull/10XX
+[#1085]: https://github.com/stackabletech/operator-rs/pull/1085
 
 ## [0.96.0] - 2025-08-25