+++ crates/stackable-operator/CHANGELOG.md
@@ -8,7 +8,14 @@ All notable changes to this project will be documented in this file.
 
 - BREAKING: Add a new CLI flag/env to disabling CRD maintenance: `--disable-crd-maintenance` ([#1085]).
 
+### Removed
+
+- BREAKING: Remove the Merge implementation for PodTemplateSpec ([#1087]).
+  It was broken because the instance was overriden by the given defaults.
+  This function is not used by the Stackable operators.
+
 [#1085]: https://github.com/stackabletech/operator-rs/pull/1085
+[#1087]: https://github.com/stackabletech/operator-rs/pull/1087
 
 ## [0.96.0] - 2025-08-25