+++ crates/stackable-operator/CHANGELOG.md
@@ -11,8 +11,12 @@ All notable changes to this project will be documented in this file.
 ### Removed
 
 - BREAKING: Remove the Merge implementation for PodTemplateSpec ([#1087]).
-  It was broken because the instance was overriden by the given defaults.
+  It was broken because the instance was overridden by the given defaults.
   This function is not used by the Stackable operators.
+
+### Fixed
+
+- Don't default the `termination_grace_period` of the `ProbeBuilder` to 0, as this is an invalid value ([#XXXX]).
 
 [#1085]: https://github.com/stackabletech/operator-rs/pull/1085
 [#1087]: https://github.com/stackabletech/operator-rs/pull/1087